feat: extract Notification event example into its own file
Moves the slack-on-idle-notification policy out of policies-advanced/index.js
into a standalone examples/policies-notification.js so it can be iterated on
independently and is easier to discover.
